separates tests for Array.
